+++ advanced-dynamic-pricing-for-woocommerce/BaseVersion/Includes/Database/Models/OrderItem.php
@@ -140,9 +140,9 @@ class OrderItem
         $sql = /** @lang MySQL */
             "CREATE TABLE {$tableName} (
             id INT NOT NULL AUTO_INCREMENT,
-            order_id INT NOT NULL,
-            order_item_id INT NOT NULL,
-            product_id INT NOT NULL,
+            order_id bigint(20) unsigned NOT NULL,
+            order_item_id bigint(20) unsigned NOT NULL,
+            product_id bigint(20) unsigned NOT NULL,
             rule_id INT NOT NULL,
             amount DECIMAL(50,2) DEFAULT 0,
             qty INT DEFAULT 0,
